refactor: draft version multiple plugins
